requirements.txt
pyjnius==1.3.0
PyYAML==5.3.1
PyMySQL==0.10.0
click==7.1.2